// src/VTodoGenerator.d.ts
import { inputObj, relatedToType, rruleType } from "./typeDefinitions";
declare class VTodoGenerator {
    due?: string;
    dtstamp?: string;
    uid?: string;
    categories?: string[];
    completed?: string;
    summary: string;
    created?: string;
    completion?: string | number;
    status?: string;
    relatedto?: string | relatedToType | relatedToType[];
    priority?: string | number;
    recurrenceid?: string;
    description?: string;
    start?: string;
    class?: string;
    rrule?: rruleType;
    geo?: string;
    location?: string;
    organizer?: string;
    sequence?: number | string;
    resources?: string | string[];
    url?: string;
    recurrences?: inputObj[];
    tz?: string;
    constructor(todoObject: inputObj);
    varNotEmpty(variable: any): boolean;
    generate(skipVCALENDAR?: boolean): string;
    getISO8601Date(date: string | number, skipTime?: boolean): string | null;
    static getValidStatusValues(): string[];
    static statusIsValid(status: string | undefined): boolean;
}
export default VTodoGenerator;